pyproject.toml
[project]
name = "socket-io"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "yokoba", email = "31529581+yokoba@users.noreply.github.com" }
]
dependencies = [
    "fastapi==0.99.0",
    "python-socketio[client]>=5.11.2",
    "uvicorn[standard]>=0.30.1",
    "rich>=13.7.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/socket_io"]
